monitor/requirements.txt
django

# PostgresSQL
psycopg2

# Rest Framework
djangorestframework==3.7.7
